fixed syntax and made default sightline grid smaller, fetch() now works
